Get rid of infer_ctxt_ext
